refactor(product): alterar tipo de estoque de Long para Integer nas classes Produto, ProdutoRequestDTO e ProdutoResponseDTO

+++ src/main/java/br/com/api/pedidos/product/dto/ProdutoRequestDTO.java
@@ -6,6 +6,6 @@ public record ProdutoRequestDTO(
         String nome,
         String descricao,
         BigDecimal preco,
-        Long estoque
+        Integer estoque
 ) {
 }
